Virada parcial do T2

+++ app/Models/Ingredient.cs
@@ -6,9 +6,8 @@ namespace TasteUfes.Models
 {
     public class Ingredient : Entity
     {
-
         [Required]
-        public int Quantity { get; set; }
+        public double Quantity { get; set; }
 
         [Required]
         public Measures QuantityUnit { get; set; }
@@ -16,7 +15,7 @@ namespace TasteUfes.Models
         [ForeignKey("FoodId")]
         public Food Food { get; set; }
         public Guid FoodId { get; set; }
-
+        
         [ForeignKey("RecipeId")]
         public Recipe Recipe { get; set; }
         public Guid RecipeId { get; set; }
